refactor: action pkgs
