+++ action/command/api.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"bytes"
