+++ action/command/check.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"log/slog"
