+++ action/command/root.go
@@ -1,4 +1,4 @@
-package main
+package command
 
 import (
 	"encoding/json"
@@ -97,5 +97,35 @@ func writeOutputJSON(w *world.World) func(cmd *cobra.Command, args []string) err
 			return errors.Wrapf(err, "failed to write output file: %s", w.Output)
 		}
 		return nil
+	}
+}
+
+func checkVersionCompatibility(client *Client, cliVersion string) {
+	if cliVersion == "unknown" {
+		slog.Warn("CLI version unknown, unable to check compatibility")
+		return
+	}
+
+	actuatorInfo, err := client.getActuatorInfo()
+	if err != nil {
+		slog.Warn("Unable to get server version for compatibility check", "error", err)
+		return
+	}
+
+	serverVersion := actuatorInfo.Version
+	if serverVersion == "" {
+		slog.Warn("Server version is empty, unable to check compatibility")
+		return
+	}
+
+	if cliVersion == "latest" {
+		slog.Warn("Using 'latest' CLI version. It is recommended to use a specific version like bytebase-action:" + serverVersion + " to match your Bytebase server version " + serverVersion)
+		return
+	}
+
+	if cliVersion != serverVersion {
+		slog.Warn("CLI version mismatch", "cliVersion", cliVersion, "serverVersion", serverVersion, "recommendation", "use bytebase-action:"+serverVersion+" to match your Bytebase server")
+	} else {
+		slog.Info("CLI version matches server version", "version", cliVersion)
 	}
 }
